Shift helper functions to tools.py and create core.py
These had to all be cleaned out of go.py so we could focus on having
that file as the server and nothing more. It’s now clearer where the
imports come from.
