+++ toplinks.html
@@ -1,7 +1,7 @@
 {% extends "base.html" %}
 
 {% from "listinc.html" import renderlink with context %}
-{% set username = getSSOUsername(False) %}
+{% set username = tools.getSSOUsername(False) %}
 
 {% block keyword %}toplinks{% endblock keyword %}
 {% block title %}Top {{ n }} go/ Links{% endblock title %}
@@ -12,7 +12,7 @@
 <h3 class="center">Top {{ n }} Links</h3>
 <div class="span12 column">
 <table class="table table-striped">
-  {% for idx, link in enumerate(byClicks(g_db.getNonFolders())[:n|int]): %}
+  {% for idx, link in enumerate(tools.byClicks(MYGLOBALS.g_db.getNonFolders())[:n|int]): %}
     {{ renderlink(idx+1, link, username) }}
   {% endfor %}
 </table>
