+++ variables.html
@@ -1,6 +1,6 @@
 
 {% extends "base.html" %}
-{% set username = getSSOUsername(False) %}
+{% set username = tools.getSSOUsername(False) %}
 
 {% block keyword %}variables{% endblock keyword %}
 {% block body %}
